Reorganize container components into a different directory
