+++ src/components/Explorer.js
@@ -2,14 +2,13 @@ import PropTypes from 'prop-types'
 import React from 'react'
 
 import AboutTheData from './AboutTheData'
-import AgencyChartContainer from './AgencyChartContainer'
-import ExplorerHeaderContainer from './ExplorerHeaderContainer'
-
-import NibrsContainer from './NibrsContainer'
+import AgencyChartContainer from '../containers/AgencyChartContainer'
+import ExplorerHeaderContainer from '../containers/ExplorerHeaderContainer'
+import NibrsContainer from '../containers/NibrsContainer'
 import NotFound from './NotFound'
-import SidebarContainer from './SidebarContainer'
-import SparklineContainer from './SparklineContainer'
-import TrendContainer from './TrendContainer'
+import SidebarContainer from '../containers/SidebarContainer'
+import SparklineContainer from '../containers/SparklineContainer'
+import TrendContainer from '../containers/TrendContainer'
 
 import { updateApp } from '../actions/composite'
 import { showTerm } from '../actions/glossary'
