+++ src/containers/App.js
@@ -3,13 +3,13 @@ import React from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 
-import BetaBanner from './BetaBanner'
-import ClearCacheBtn from './ClearCacheBtn'
-import Disclaimer from './Disclaimer'
-import Feedback from './Feedback'
-import Footer from './Footer'
-import Glossary from './Glossary'
-import Header from './Header'
+import BetaBanner from '../components/BetaBanner'
+import ClearCacheBtn from '../components/ClearCacheBtn'
+import Disclaimer from '../components/Disclaimer'
+import Feedback from '../components/Feedback'
+import Footer from '../components/Footer'
+import Glossary from '../components/Glossary'
+import Header from '../components/Header'
 import * as feedbackActions from '../actions/feedback'
 import * as glossaryActions from '../actions/glossary'
