+++ src/containers/ExplorerHeaderContainer.js
@@ -2,10 +2,10 @@ import startCase from 'lodash.startcase'
 import React from 'react'
 import { connect } from 'react-redux'
 
-import ExplorerIntro from './ExplorerIntro'
-import Loading from './Loading'
-import PlaceThumbnail from './PlaceThumbnail'
-import UcrResourcesList from './UcrResourcesList'
+import ExplorerIntro from '../components/ExplorerIntro'
+import Loading from '../components/Loading'
+import PlaceThumbnail from '../components/PlaceThumbnail'
+import UcrResourcesList from '../components/UcrResourcesList'
 import { getPlaceInfo } from '../util/place'
 import { getAgency, oriToState } from '../util/ori'
