+++ src/containers/NibrsContainer.js
@@ -4,10 +4,10 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { connect } from 'react-redux'
 
-import ErrorCard from './ErrorCard'
-import Loading from './Loading'
-import NibrsCard from './NibrsCard'
-import Term from './Term'
+import ErrorCard from '../components/ErrorCard'
+import Loading from '../components/Loading'
+import NibrsCard from '../components/NibrsCard'
+import Term from '../components/Term'
 import parseNibrs from '../util/nibrs'
 import { getAgency, oriToState } from '../util/ori'
 import { getPlaceInfo } from '../util/place'
