+++ src/containers/SidebarContainer.js
@@ -2,9 +2,9 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { connect } from 'react-redux'
 
-import CrimeTypeFilter from './CrimeTypeFilter'
-import LocationFilter from './LocationFilter'
-import TimePeriodFilter from './TimePeriodFilter'
+import CrimeTypeFilter from '../components/CrimeTypeFilter'
+import LocationFilter from '../components/LocationFilter'
+import TimePeriodFilter from '../components/TimePeriodFilter'
 import { hideSidebar } from '../actions/sidebar'
 import { getAgency, oriToState } from '../util/ori'
 import { nationalKey } from '../util/usa'
