+++ src/containers/TrendContainer.js
@@ -5,12 +5,12 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { connect } from 'react-redux'
 
-import DownloadDataBtn from './DownloadDataBtn'
-import ErrorCard from './ErrorCard'
-import Loading from './Loading'
-import NoData from './NoData'
-import TrendChart from './TrendChart'
-import TrendSourceText from './TrendSourceText'
+import DownloadDataBtn from '../components/DownloadDataBtn'
+import ErrorCard from '../components/ErrorCard'
+import Loading from '../components/Loading'
+import NoData from '../components/NoData'
+import TrendChart from '../components/TrendChart'
+import TrendSourceText from '../components/TrendSourceText'
 import { generateCrimeReadme } from '../util/content'
 import { getPlaceInfo } from '../util/place'
 import mungeSummaryData from '../util/summary'
